+++ backend/Popcorn_Prophet/src/main/java/com/popcorn_prophet/popcorn_prophet/service/ProductReviewService.java
@@ -1,6 +1,8 @@
 package com.popcorn_prophet.popcorn_prophet.service;
 
 import com.popcorn_prophet.popcorn_prophet.DTO.ProductReviewCreateDTO;
+import com.popcorn_prophet.popcorn_prophet.entity.Member;
+import com.popcorn_prophet.popcorn_prophet.entity.Product;
 import com.popcorn_prophet.popcorn_prophet.entity.ProductReview;
 import com.popcorn_prophet.popcorn_prophet.repo.ProductReviewRepository;
 import lombok.RequiredArgsConstructor;
@@ -10,6 +12,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,32 +22,52 @@ public class ProductReviewService {
     private final ProductService productService;
     private final MemberService memberService;
 
-    public Page<ProductReview> getProductReviews(Long productId,int page) {
+    public Page<ProductReview> getProductReviews(Long productId, int page) {
         return productReviewRepository.findAllByProductId(productId, PageRequest.of(page, 5));
 
     }
 
 
-
-    public ProductReview addProductReview(ProductReviewCreateDTO productReview, Long productId, Long memberId) {
+    public Optional<ProductReview> addProductReview(ProductReviewCreateDTO productReview, Long productId, Long memberId) {
+        Optional<Product> product = productService.getProduct(productId);
+        Optional<Member> member = memberService.getMember(memberId);
+        if (product.isEmpty() || member.isEmpty()) {
+            return Optional.empty();
+        }
         ProductReview newProductReview = ProductReview.builder().review(productReview.getReview()).rating(productReview.getRating()).
-                product(productService.getProduct(productId)).member(memberService.getMember(memberId)).build();
+                product(product.get()).member(member.get()).build();
 
-        return productReviewRepository.save(newProductReview);
+        return Optional.of(productReviewRepository.save(newProductReview));
     }
 
-    public void deleteProductReview(Long productReviewId) {
+    public Optional<Boolean> deleteProductReview(Long productReviewId) {
+        Optional<ProductReview> productReview = productReviewRepository.findById(productReviewId);
+        if (productReview.isEmpty()) {
+            return Optional.empty();
+        }
         productReviewRepository.deleteById(productReviewId);
+        return Optional.of(true);
     }
 
-    public ProductReview getProductReview(Long productReviewId) {
-        return productReviewRepository.findById(productReviewId).get();
+    public Optional<ProductReview> getProductReview(Long productReviewId) {
+        Optional<ProductReview> productReview = productReviewRepository.findById(productReviewId);
+        if (productReview.isEmpty()) {
+            return Optional.empty();
+        }
+        return productReview;
     }
 
-    public ProductReview updateProductReview(ProductReview productReview) {
-        return productReviewRepository.save(productReview);
-    }
+    public Optional<ProductReview> updateProductReview(ProductReview productReview) {
+        Optional<ProductReview> productReviewOptional = productReviewRepository.findById(productReview.getId());
+        if (productReviewOptional.isEmpty()) {
+            return Optional.empty();
+        }
+        ProductReview productReviewToUpdate = productReviewOptional.get();
 
+        productReviewToUpdate.setRating(productReview.getRating());
+        productReviewToUpdate.setReview(productReview.getReview());
+        return Optional.of(productReviewRepository.save(productReviewToUpdate));
+    }
 
 
 }
